- Added error handling to RestClient and renamed the connector.

+++ TraderTerminal/Trader/Models/Ticker.cs
@@ -3,14 +3,14 @@ namespace Trader.Models;
 public class Ticker
 {
     public string Pair { get; set; }
-    public float Bid { get; set; }
-    public float BidSize { get; set; }
-    public float Ask { get; set; }
-    public float AskSize { get; set; }
-    public float DailyChange { get; set; }
-    public float DailyChangeRelative { get; set; }
-    public float LastPrice { get; set; }
-    public float Volume { get; set; }
-    public float High { get; set; }
-    public float Low { get; set; }
+    public decimal Bid { get; set; }
+    public decimal BidSize { get; set; }
+    public decimal Ask { get; set; }
+    public decimal AskSize { get; set; }
+    public decimal DailyChange { get; set; }
+    public decimal DailyChangeRelative { get; set; }
+    public decimal LastPrice { get; set; }
+    public decimal Volume { get; set; }
+    public decimal High { get; set; }
+    public decimal Low { get; set; }
 }
